+++ src/TagBites.Expressions/Expressions/Extensions/DelayThrowExpression.cs
@@ -1,0 +1,8 @@
+using System.Linq.Expressions;
+
+namespace TagBites.Expressions.Extensions;
+
+internal class DelayThrowExpression(Expression exception) : Expression
+{
+    public Expression Exception { get; } = exception;
+}